Add support for the stm32f769disco board.
This also improves the handling of the stm32f746 disco board by properly
selecting the cortex-m7 as cpu as well as selecting the proper fpu.

Regenerate the SVD binding for the stm32 boards, and add the stm32f769
bindings.

Refactor a bit i-sbbbopa.ads so that the FLASH latency is now included, and
the Clock_Frequency identifier never clashes with i-sbbpara.ads.

TN: P721-006

git-svn-id: svn+ssh://svn.us.adacore.com/Dev/trunk/bb-runtimes@323434 f8352e7e-cb20-0410-8ce7-b5d9e71c585c