Move the install script from the embedded-runtimes
Since bb-runtimes is now public and the entry point for BSPs. This is
the right place to have the script to install the BSP.

The only modification to the embedded-runtimes version of install.py is
the call to 'build-rts.py --bsps-only' to generate the BSPs dir before
compiling and installing the run-times.

Change-Id: Ia772b2cb8f6349c572052fe0d7fcc8aa8356569b