ravenscar: improve 64-bit timers support.
This change use a different spec for s-bbbosu: the Set_Alarm subprogram
now takes an absolute time as parameter, so a 64-bit value.

On Aarch64 and cortex-a platforms, we use this absolute value to set the
physical clock comparator, thus enabling long delays, and removing a slight
bias introduced by the diff between clock and alarm.

On the other platforms, only relative delays are available:
* on PPC, we have to stick with the old behavior, as the maximum delay
  available is 16#7FFF_FFFF# we raise periodically timer interrupts until
  the final alarm time is reached.
* on TMS570, we set the delay in two steps: first a long coarse delay to
  reach the point where delta time is less than 2**33. Once the diff is
  small, we set a second delay to reach exactly the Alarm time.

Change-Id: I9f49bb1df7e31b9a421249446de435f4405efe1d
TN: Q630-016