TMS570: continued work at integrating support for TMS570LC43.
This change simplifies the startup codes by using a single startup
that applies to the various loaders.

This also brings separate clock/peripherals startup for the LS31 and the
LC43

Finally, the SCI I/O now takes into account the differences in VCLK values
between the two MCUs.

TN: Q726-004