TMS570: fix boot loader for Cortex-R5
This moves most of the board init from crt0.S to Ada code, for readability
and maintenance purpose.

This also refines the MPU tables to take into account the various
memory layouts.

Finally, this moves the hardware workaround to separate files depending
on whether the CPU is cortex-r4 or cortex-r5.

TN: Q711-032 and Q726-004