TMS570: fix typo in MPU regions initialisation, and boot on qemu.
This fixes a type in MPU regions that prevented one of the peripheral
regions to be properly set.

Also use the MPUIR register to retrieve the total number of available MPU
regions. This ensure that only those are initialized.

Qemu sets the reset condition to 0, similarly to what an external bootloader
would do. In such case, don't initialize the MPU and cache.

TN: QA06-053