remove unnecessary comment from system-xi-sparc-ravenscar.ads
This makes this system.ads file consistent with the other system.ads
files, and in particular allows us to get the exact same output
on all baremetal targets when running the lxh4005 ACATS test using
a ravenscar-sfp runtime (same line number flagging the restriction
causing the build error).

For R209-017.