leon3-uart: fix the tx_ready check
Is_Tx_Ready check is broken for implementation without FIFO because the
FIFO_Full bit is not implemented in this case.

So let's add a check which works in both case and keep the advantage of the
FIFO when it is implemented.

TN: R224-005
Change-Id: I2c1a3e91e6cc949bf599fc5434e06ecab435280a
(cherry picked from commit 50614b02ca40582b510022f0cbd40cbbea05dcc0)