arm ld scripts: move __exidx_start/end within the .exidx sections
For _start, important to make sure that the symbol address coincides
with the start of section, avoiding distance possibly created out of
the section start alignment.

For _end, just to be consistent with _start.

In the course of QA05-001 (ravenscar c_rpf09_15 fails on arm-pikeos),
where we discovered that most arm bareboard ports were affected as well
as pikeos.

Change-Id: I3158f00aa356beb6bff68c7f12ebc60e8b7f5f41