HiFive1: Fix .srodata and .sdata LMA alignment
Sections' VMA and LMA can have a different alignment.

Since we load all three sections (.data, .srodata and .sdata) as a
single block, we expect them to be contiguous in ROM as well as in RAM.
But a mismatch in VMA/LMA alignment means that 2nd and 3rd sections can
have padding in their RAM address but not in their load address (ROM).
The result is that a section can be loaded at an incorrect address.

For example here:

.data    0x0000000080000000   0xb load address 0x0000000020401ff0
.srodata 0x000000008000000b   0x0 load address 0x0000000020401ffb
.sdata   0x000000008000000c  0xf0 load address 0x0000000020401ffb

.sdata is aligned in RAM but not in ROM. The linker expects .sdata to be
copied at 0x8000000c but the startup code will copy it at 0x8000000b.

To fix this problem we explicitly ask the linker to align both the VMA
and LMA of .srodata and .sdata.

Part of R326-058.

Change-Id: I8ceabd3a7d41c7f843c7cdf650463c0fb10e1aa1