Add support for runtime sources generated from template files.
This adds the ability to substitute certain keywords during
generation of the runtime source trees, based on a template file.
Instances of ${var} in the source file are replaced with a value
specified by the runtime being generated.

The primary purpose of this is to allow for customizing the
interrupt priority levels in System.ads for ARM Cortex-M devices,
where the number of interrupt priorities may vary for different
Cortex-M devices. For example, STM32 have 16 interrupt priorities,
but nRF52 only have 8. system-xi-cortexm4-full.ads uses the
variables ${Max_Priority} and ${Max_Interrupt_Priority} in the
places in the file where those values are required, and those
variables are replaced with the correct value when the source
tree is generated.