riscv: common-RAM.ld: discard empty sections
". = ALIGN(0x4)" will output sections even when they are empty as explained
in the ld manual.  This is not good: it can confuse the linker in certain
circumstances.

Let's replace them with ". = ALIGN(. != 0 ? 0x4 : 0x1)" which will emit the
section only when it is not empty (the padding won't emit the section
either).

With this change the comment which requires the end of the .data section to
be word aligned in order to be loaded properly is statisfied, and the top
of both .rodata and .srodata is word aligned as well which might be not
required but this keep the previous behavior.

TN: T226-028
Change-Id: Ib1b370a581da74f21c96d71798d5bf6d89c53231