x86_64-elf-multiboot: Place all code above 1MB
To work around QEMU not allowing loading of elf64-x86-64 images using the
-kernel switch, we load the ELF file as a contiguous blob. However, some
versions of QEMU don't like us loading zeros over the BIOS region. So for
now, relocate the startup code above the bios region. To provide multicore
support, this will have to be reversed as cores boot in 16-bit real mode.
For that case it will be easier to drop the use of the -kernel switch and
switch to booting via grub2 using multiboot2.

Change-Id: Ia7e13fcf8be4886b5dc6a41b2192636494615608
TN: T422-004