Simplify clock setup.
This combines the two setup_pll.adb files (which are almost identical)
into a single common file to avoid excessive duplication.
Simple_Clock_Tree is defined in s-bbmcpa.ads and is used to determine
which clock features are available in setup_pll.adb so that invalid
clock configs can be rejected. For example, requesting the HSI48 clock
on the STM3203x is rejected as HSI48 is not available on that device.

This also moves the HSE_Enabled variable into setup_pll.adb where it
is calculated automatically based on the requested clock sources.
The HSE is only enabled if it is requested for system clock generation
(either directly used to generate SYSCLK, or is used indirectly as
the PLL source which is then used to generate SYSCLK).