ppc-elf: adjust FRAME_SIZE in handler.S to keep the stack 16-bytes aligned
The PowerPC ELF ABI requires the stack pointer to maintain a 16-byte
alignment. For our interrupt handlers we were miss-aligning the stack
pointer by 8-bytes when we allocated space to save the interrupted context.
In practise, the PowerPC architecture can handle 8-byte stack alignment (as
demonstrated by the EABI) which is why we haven't had any issues here.

Change-Id: Ie33e8f888dbbfb6a6ea3fe0474d3fa9e3c479b29
TN: TA22-033