RISCV64: Add GDB readable value that specifies the base HW CPU id
This is because of the specific architecture of the PolarFire SOC (and
Sifive Unleashed) where HW CPU #0 is actually not used by the run-time
because it has a different instruction set (power monitor CPU).

For all other architectures/boards:
 - Ada CPU #1 is HW CPU #0
 - Ada CPU #2 is HW CPU #1
 - etc.

On the PolarFire SOC:
 - Ada CPU #1 is HW CPU #1

This is not even RISC-V specific, it is board specific.

GDB needs to be able to identify which CPU a Ravenscar task is running
on to resume that CPU execution when user wants to resume a task
execution. With this discrepancy in the CPU ids, GDB is unable to do
that.

This value will tell GDB which HW CPU corresponds to the first Ada CPU
id.

Part of TA27-025.

Change-Id: Ic69cbe2ea6d24bb881b727bb3f6ecae1370c843f