Implement customizable runtimes
This adds a mechanism to allow user-defined board parameters to
configure certain parts of the runtime during generation.
For example, to configure the system clocks.

Configurations are specified in JSON files which are loaded
by build_rts.py and define board parameters that are used by
the Python code to parameterize the runtime source files.
This makes use of the template system to pass values from the
JSON file into the runtime source files.

Currently, only stm32f0xx and nRF52 targets can be customized
with this new mechanism. Other targets are not yet configurable.