update following renaming of some System.GCC units
Some System.GCC units were renamed to be children of System.GCC.DI
instead, so as to have System.GCC be free of symbols which are
specific to 32bit platforms. This was needed in the context
of a 64bit vx7r2cert platform (AArch64) where we needed to add
equivalent routines specific to 64bit platforms ("ti" functions
instead of "di" ones). Those routines were added under System.GCC.TI.

This commit adjusts the sources and documentation in this repository
accordingly.

Change-Id: Ib072d833f62335d0aa600e50d8710fa8123e4bfe
TN: TB25-032