PowerPC runtimes: don't use r0 to load initial stack pointer in start-ram.S
For the _start_ram routines for PowerPC, the program stack pointer is loaded
into a temporary register in order to backlink to the current stack pointer.
However, r0 shouldn't be used for this temporary register as it is treated
as the value zero when used as the store destination register (which happens
when we store the backlink).

Also better document that the runtime will return to the bootloader if the
program terminates.

Change-Id: I55f945de19c4d3780096dde54449d63a477cf7ac
TN: TC17-014