Make sure we don't use FPU registers in the ravenscar kernel.
This can have unexpected results with regard to the FPU registers, if they
are modified in critical sections of the ravenscar runtime.

FPU registers can be used when initialising structures using FPU registers
(because they are larger, so faster), or when vectorizing the code with SIMD
instructions.

This change:
* builds now two runtime libraries for the ravenscar-full, so that separate
  set of switches can be used for libgnarl
* introduce libgnarl-specific switches
* uses -mgeneral-regs-only when building libgnarl to ensure no FPU register
  is used.

Change-Id: I9c12de1c671b76c3b7428ecf4d74609dc4dffbc1
TN: TA30-022