libgnarl is now needed to link with the ravenscar-full runtime.
The whole libgnarl used to be merged with libgnat, but since TA30-022,
this is not the case anymore. But with ravenscar-full, there's still
interdependencies between libgnat and libgnarl (due to the way softlinks
are implemented on bare metal). This change fixes this by forcing libgnarl
at link time in runtime.xml.

Change-Id: I7bb56f2986e91f2b20e3a96dbe7c355c010a13a9
TN: U112-036