Refine the ARM machine reset implementation.
Apply a barrier prior to the reset request to ensure previous
memory accesses complete before the reset request occurs.

Disable interrupts prior to the reset. Depending on the implementation,
the reset could take some time to occur, during which an interrupt could
come in. In that case the reset could occur in the middle of the
interrupt handler.

Add comment explaining the presence of the null infinite loop following
the reset request.

TN: U325-048
Change-Id: I4dc05ee7592511ce61294b729a0641d642a45b3c