Select runtime format by specific selector instead of pikeos
The old runtime format is not only needed for pikeos but
also for vx7r2cert. Since setting is_pikeos on vx7r2cert
doesn't make sense the selection of the runtime format is
now done by the is_legacy_format property. The default
implementation has the same behaviour as the is_pikeos
property but it can be overwritten.

TN: U518-028
Change-Id: I56f317b52735bd451a6a97b46714d3b7036e7b13