Fix aarch64 zynqmp mmu implementation
This patch reenables the MMU for the aarch64 zynqmp board.
The fix for the initial problems was to replace the raised
Program_Error with a call to System.Machine_Reset.Stop.
The problem was that on ravenscar-full the last chance handler
creates a circular symbol dependency. Furthermore it depends on
the secondary stack that is not set up when the MMU is initialized.

TN: U423-022 U611-041
Change-Id: I211cfc8410e8b5d6c7f11a8f1fe12e87dc489c63