Fix Leon3 FPU context switch
The lazy FPU context switch mechanism requires the FPU to be disabled at
the beginning of a task. For the secondary tasks, the FPU is disabled in
the FSR value of their context buffer. But the for environment task it
is not disabled after the crt0 initialization.

This patch disables the FPU for the environment task during elaboration,
in the Initialize_CPU procedure.

The FPU context switch is also slightly modified to always load a
context from a buffer, even when there is no previous FPU user
registered. This means that now the FPU context is always loaded from
task buffer instead of inherited from the crt0 initialization for one of
the task (the first one to use the FPU).

Part of U727-033.

Change-Id: I3bf167412a07a4fc18dca6c022ae5b0c79c083b0