Fix FMA and FPU handling
Previously the properties has_single_precision_fpu
and has_double_precision_fpu indicated if floating point
code should be used or not. Additionally FMA support was enabled
by architecture. Due to the change in the runtime profile
floating point support is now required for all platforms. This
also indicated that FMA was incorrectly added when no hard float
support was available. FMA is now only used if an actual FPU is
available. Additionaly the target for the HiFive1 board was fixed
that previously incorrectly stated to have an FPU.

TN: U818-028 U903-028
Change-Id: I9f0e312742066a6cdf552740f0423f80f56b0432