Adjust x86_64-elf runtime so that it runs without a BIOS on QEMU
Overwriting the BIOS QEMU loads is problematic as QEMU does not define if
the program image or BIOS is loaded into memory first. Consequently, the
x86_64-elf runtime is adapted to work without a BIOS by loading a long jump
instruction to the program's boot code at the processor's boot address.

Change-Id: I8a454e3f4010720e466b30a62c3e9c2a2f3f62b5
TN: UA12-047