Adjust our way to link libc with embedded runtimes
We were adding -lc explicitly without telling gcc not to do it
and we didn't quite account for all the possible intricacies
in dependencies between libc, libgnat and libgcc.

Pass -nolibc to inhibit gcc's processing (preventing
a potentially problematic inclusion of libgloss in some
configurations), and resort to --start-group/--end-group
to robustify the libc/libgnat/libgcc processing.

Part of UA18-004 (riscv64-elf transition to gcc-11)

Change-Id: I2c54c491db2bd411899e0e9e6696f02125d7f21f