Fix zynq boot sequence when MMU is pre-configured.
In particular with bootloaders like uboot, the MMU can be configured
before the Ada application starts. Setting the DACR in this context
may lead to an abort, with the code not being in an executable page
anymore.

Also fix mmu table being added twice in the runtime (so one is useless).

Change-Id: I35e3d17cd2d5ada27523f70a16192ce8f5c69e96
TN: U813-019