Zynq7k: fix issue with nested interrupts
Previously, the user irq handler was executed in the CPU IRQ mode.
So if the handler was interrupted by a higher priority interrupt,
the LR[IRQ] register could end up being corrupted in the first
handler, and thus lead to an invalid execution flow.

This fix changes the execution mode, and user handler now uses the
SVC mode to execute, allowing proper banked LR registers in case of
nested interrupts. This is possible because the runtime does not
use this mode otherwise.

Change-Id: I29c6f31a7910b49a6ec04d37bae3d3c16abb42f1
TN: V114-047