Better fix for V114-047
This moves the low level IRQ trap to a separate asm source, so that it's
not included with the light runtime, renames s-bbcppr to show that
the difference between the two ARM ones is whether or not a GIC is used.

Adjusted the zynqmpr5 runtime to use the GIC variant.

Change-Id: I8fb4f4235acf91377bccda73ea6e2264bb0b8acd