Save runtime version in rts.json
The runtime version is parsed from gnatvsn.ads in the gnat dir.
The path of this file is only passed to gen_rts_sources.py and
therefor only available when the sources are generated. Previously
it was parsed when build_rts.py is called however this fails in
some configurations. To prevent this the version is now encoded
in rts.json when the sources are generated and then read from that
file when the runtime is built.

TN: V524-041
Change-Id: Id6394a5c6fba83a7d356b7bc512659c5d3e2d3e8