Deos: Update compiler switches, System and use cert version of s-secsta
Runtime needs to be compiled with -fPIC. Switches required to build a Deos
application moves to runtime.xml rather than using the Deos makefiles as
they only appear in the Deos make file when the target has C++ code.

System is adjusted so the main procedure is called main and the cert version
of System.Secondary_Stack is used since soft-links is used by cert.

Change-Id: If682bca1c73dc3560350745ab49d9f812618429b
TN: V712-053