Cert: don't use Arm version of System.Exceptions.Machine on Arm platforms
For certification applications, we use SJLJ exceptions as it is viewed that
ZCX exceptions are too difficult to certify. However, the Arm EABI defines
the ZCX as the exception scheme. Consequently for the Cert runtime, don't
use the Arm implementation of System.Exceptions.Machine since were breaking
from the Arm EABI here and instead use the normal GCC version.

Change-Id: Ib3287ff5750e9cb3b8e040cbb2f4c856d791e4e3
TN: V712-053