Compile applications with -fno-tree-loop-distribute-patterns
The tree-loop-distribute-patterns optimization looks for code patterns that
can be replaced with library calls, for example memset and strlen. This
creates a hidden runtime dependency that is considered undesirable by many
of our customers. The recognition of strlen and rawmemchr like loops is
particularly problematic, as these functions may not be available on certain
bare-metal or cert runtimes.

Also make each instance of the compiler_switches property return a tuple.

Change-Id: I62b1456e3668f38927d96d37ab95e8ce5be5243e
TN: VA23-004