Enable Support_Atomic_Primitives on arm and ppc
The addition of a 32-bit version of s-atopri allows
us to enable Support_Atomic_Primitives on Arm Cortex-M
and 32-bit PPC platforms. Additionally, the atomic
packages are now available on all runtime profiles
as they are needed to provide a thread safe
initialization of the secondary stack. This also
applies to the light runtime as it may run in a
context with foreign threads.

TN: V823-036
Change-Id: Ic88094521090e70e61c12ba8b0c16b4b3becbc3d