Select version for shared libgnat from Library_Version
The version of the shared libgnat was set with the same
mechanism as the GNAT version. However while the GNAT version
always contains a date the library version for release
compilers does not contain a date. To keep the runtime in
sync with GNAT the Library_Version in the runtime projects
is now derived from Library_Version in gnatvsn.ads instead of
GNAT_Static_Version_String.

TN: W403-017