spike: move htif symbols to 0x100000
The htif section defined two symbols retrieved by the spike simulator or
an emulator to communicate more easily with the host.
We used to provide these symbols at the address 0x0, but qemu-8 is now
forbidding it.
Improve the Ada part to retrieve these symbols' address through the
linker script.

TN: W515-007