Place .data.rel.ro.local sections in .data.rel.ro on morello-elf
The linker script wasn't placing .data.rel.ro.local in the .data.rel.ro
section, so the linker was instead placing it where it saw fit.

In a purecap environment the .data.rel.ro.local section needs to be
placed before the .got so that it is within the bounds of the
program counter capability (PCC). Otherwise, PCC-relative loads
would fail at runtime due to a capability bounds fault.

Issue: eng/toolchain/bb-runtimes#1