x86_64-elf: ensure unwinder terminates in interrupt stack
In the interrupt and exception handlers, an explicit null return address
needs to be pushed onto the stack prior to calling any Ada procedures to
inform the unwinder that it has reached the end of the stack. This is needed
as the handlers are naked functions with no stack frame.

The formatting of the Fatal_Exception messages has also been cleaned up.

Closes eng/toolchain/gcc#5