morello-elf: Fix stack pointer corruption when handling FP traps
The post-increment version of the ldp instruction doesn't preserve
the capability metadata when it is used to update the stack pointer
capability, resulting in an invalid capability that cannot be
dereferenced. The workaround is to use a separate 'add' instruction
to update the stack pointer.

Part of Issue: eng/toolchain/bb-runtimes#6