Fix shared runtime builds
The shared build of libgnarl sometimes failed due to incorrect
linker flags. The reason was the setting of the required flags
in the package Builder. Flags set in this package are ignored
if more than one language is used in a library project which is
the case for libgnarl. The flags are now set properly via
Library_Options.

ref eng/cert/certified-rts#9