Split morello runtimes into two variants
This splits the morello runtime into two variants: "morello" and
"morello-semihosting". The two runtimes are identical except for the
implementation of Ada.Text_IO and System.Machine_Reset. The "morello"
runtime uses a serial port for text I/O and a machine reset that sits
in a loop, whereas the "morello-semihosting" runtime uses a semihosting
implementation.

Issue: eng/toolchain/bb-runtimes#15