Use libunwind for exception handling with GNAT-LLVM on ZynqMP
The LLVM library libunwind is a replacement for the parts of libgcc that
perform stack unwinding and exception propagation. It uses slightly
different symbols to locate exception data in memory, which this commit
adds to the ZynqMP linker script. We also enable it by default when the
runtime is assembled for use with GNAT-LLVM.

Issue: eng/toolchain/bb-runtimes#20