Add Compiler_Backend scenario
Some runtime files depend on the compiler backend used, for example GCC
requires us to include gcc specific unwinder code. To allow us to specify
code specific to a compiler backend, a new Compiler_Backend scenario is
added.

Also added a connivence function to tell if we are building for LLVM or not
to improve code readability.

Issue: eng/toolchain/bb-runtimes#20