morello: Set DDC to null after startup.
Since we operate in purecap mode, the DDC is not needed once the various
system capabilities have been minted during startup. The DDC is therefore
set to null to prevent it from being misused by software to gain access to
objects it has not been explicitly given access to.

Issue: eng/toolchain/bb-runtimes#34