powerpc/mpc8641: add memory segments in qemu-rom.ld
These MEMORY values will help creating a common linker script called by
both the ROM and the RAM setup.

For now, qemu-rom.ld is simply adapted to take its addresses from a
local memory map.

The new .vector section aims to reserve the space for the exception
vectors. It's being moved at the before .text for the later case when
REGION_DATA and ROMABLE_REGION will be the same.

bb-runtimes#37