am64xr5: change DDR MPU settings to Normal, Cacheable
The default MPU memory map of the am64xr5 runtime marks the DDR memory as
non-sharable device memory. However device memory is designed for
memory-mapped peripherals, so doesn't allow execution even if we explicitly
mark it as execute enabled.

To be more customer friendly, change this memory region to be cacheable
normal memory (which really it is).

Issue: eng/toolchain/bb-runtimes#53