Add support for building a particular profile for a target
Add a --profile switch to build_rts.py to allow us to specify which profile
to build for a target. By default, the script will create and build runtimes
for all supported profiles. The new --profile switch allows us to create
just the runtime that is being requested by a customer.

Issue: eng/toolchain/bb-runtimes#73