Set Exit_Status_Supported to True on native targets
The light runtime for native targets was reusing the system.ads
of their respective bare metal targets. On bare metal systems
setting an exit status is not supported, hence it was not done
by the binder. On native targets this causes the program to exit
with a random exit status. This change introduces a separate
system.ads for native targets with the correct configuration.

ref eng/toolchain/bb-runtimes#76