Restore gccmaths on vx7r2cert except on x86_64
When we started testing against helix-cert (on x86_64), we saw that
its cert GOS was providing gccmaths routines; so we figured that we
would not want to provide ours, otherwise we would have a conflict at
link time.  This was implemented by eng/toolchain/bb-runtimes!94.
However, this turned out to be a peculiarity of the x86_64 platform
(and possibly a bug there); aarch64 helix-cert GOS does not provide
gccmaths (neither the other architecture over vx7r2 safety profile
like ppc*, for what it's worth).  All in all, the proper solution at
this point is to provide gccmaths on all vx7r2cert platforms except
x86_64.

eng/toolchain/acats#21