Remove invalid assertion from stm32f4 runtime setup
Remove an assertion from setup_pll.adb in the stm32f4 runtimes. That
assertion introduces a dependency to System.Assertions which does not
conform to the pragma No_Elaboration_Code_All required by setup_pll.ads.

Also fix the check itself. The assertion was inverted, so it would
always fail on successful execution.

ref eng/toolchain/bb-runtimes#85