Remove -pthread from shared library build
Remove -pthread from the linker options used to build a shared runtime.
While this switch may be needed on some systems it is not available on
QNX and causes the build to fail there. For future platforms that need
this switch a feature to set platform specific linker options will be
introduced.

ref eng/toolchain/bb-runtimes#106