ppc-lynx178: move to ada math functions and disable FPU
When enabling fpu or using __cert files, sqrt and sqrtf are replaced
by `__builtins_sqrt[f]` functions, leaving the compiler decides if
FPU instructions could be generated.

For ppc-lynx, FPU is not enabled, thus it results in calls to libm,
which in practice is fine. However, LynxOS libm doesn't define `sqrtf`.

To avoid any issues with `__builtins` are calling back missing functions
of libm, disable FPU and replace the files pulling the system features
to pure Ada implementation.

llvm-project#21