Add dependency on shared last chance handler for QNX
For QNX, add an explicit dependency from the shared Ada
run-time libraries to a library containing the last chance handler,
so that the last chance handler is always dynamically loaded
whenever the Ada runtime is loaded.

Create a dummy library with the expected name so that we can build
the runtime.

ref eng/cert/projects/c-nvidia-qnxlight2#25