Add support for C++ ctors/dtors in Ada runtime
Linker scripts include the compiler-generated sections pointing
to constructors and destructors.

Startup routines call the __libc_init_array/__libc_fini_array
subprograms to execute constructors and destructors.

The runtime.xml files direct the linker to add the crtbegin.o and
crtend.o files (via a spec file) that contain the markers required
for the lists of constructors and destructors.

For light and light-tasking bare-metal runtimes (excluding morello
that does not support C++) add the System.Libc Ada package that
mimics the functionality from the C library to execute dynamic
initialization and termination for non-local variables and routines
to execute at program termination.

ref eng/toolchain/gnat#991