Properly calculate undef handler return address on armv7r
The return address of the undefined instruction exception handler on
ARMv7R needs to rerun the last instruction after returning. For that it
needs to subtract the instructions size from the link register. This
size can be either 4 or two bit depending if the processor was in ARM or
Thumb mode at the time the exception is raised. Check the T bit of the
SPSR and calculate the correct offset.

ref eng/cert/projects/c-arianegroup-thumb#2